# api-documentation.yml
openapi: 3.0.3
info:
  title: TokenTrust API
  description: |
    AI-powered multi-source intelligence platform for Solana token fraud detection.
    
    ## Current Implementation Status
    
    ### ✅ Phase 1 Complete (Foundation & Basic Analysis)
    - Core token analysis engine
    - Helius API integration for on-chain data
    - Basic risk scoring algorithm
    - Whitelisting system for known tokens
    - Mint authority and freeze authority detection
    - Transaction volume analysis
    - Token age calculation
    - Basic metadata quality assessment
    
    ### 🚧 Phase 2 In Progress (Multi-Source Intelligence)
    - DEX integration (DexScreener, Birdeye, Jupiter)
    - Holder distribution analysis
    - Liquidity risk assessment
    - Volume pattern analysis
    - Rug pull detection
    
    ### 📋 Phase 3 Planned (AI Intelligence & Real-Time Monitoring)
    - Advanced LLM analysis (GPT-4, Claude)
    - Real-time monitoring system
    - Predictive modeling
    - Alert and notification system
    - Social media intelligence
    - Web presence analysis
    
  version: 1.0.0
  contact:
    name: TokenTrust Team
    email: support@tokentrust.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.tokentrust.ai
    description: Production server
...
paths:
...
  /api/analyze:
    post:
      tags:
        - Analysis
      summary: Analyze a Solana token for fraud risk
      description: |
        Performs comprehensive token analysis including:
        - On-chain data analysis (mint authority, freeze authority, transactions)
        - Risk scoring based on multiple factors
        - Whitelist verification
        - Transaction pattern analysis
        - Creator behavior assessment
        
        **Current Implementation**: ✅ Complete
        - Basic risk scoring with 6 factors
        - Transaction analysis
        - Creator behavior patterns
        - Whitelist system integration
        
        **Processing Time**: ~2-5 seconds
        **Accuracy**: 100% for whitelisted tokens, 70%+ for unknown tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tokenAddress
              properties:
                tokenAddress:
                  type: string
                  description: Solana token address (mint address)
                  example: "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"
                includeDex:
                  type: boolean
                  description: Include DEX and liquidity analysis (Phase 2)
                  default: false
                includeSocial:
                  type: boolean
                  description: Include social media analysis (Phase 3)
                  default: false
                includeWeb:
                  type: boolean
                  description: Include web presence analysis (Phase 3)
                  default: false
                includeAI:
                  type: boolean
                  description: Include AI-powered analysis (Phase 3)
                  default: false
      responses:
        '200':
          description: Analysis completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/TokenAnalysis'
                  processingTime:
                    type: number
                    description: Analysis time in milliseconds
                    example: 2340
                  timestamp:
                    type: string
                    format: date-time
                    example: "2024-01-15T10:30:00.000Z"
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
        '500':
          description: Analysis failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========================================
  # TOKEN DATA ENDPOINTS
  # ========================================
  /api/token/{address}/metadata:
    get:
      tags:
        - Token Data
      summary: Get token metadata
      description: |
        Retrieves comprehensive token metadata using Helius DAS (Digital Asset Standard).
        
        **Current Implementation**: ✅ Complete
        - Token name, symbol, decimals
        - Supply information
        - Price data (if available)
        - Image and description
        - Interface type
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
          description: Solana token address
          example: "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"
      responses:
        '200':
          description: Token metadata retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/TokenMetadata'
                  timestamp:
                    type: string
                    format: date-time
        '404':
          description: Token not found
        '500':
          description: Failed to fetch metadata
  /api/token/{address}/transactions:
    get:
      tags:
        - Token Data
      summary: Get token transaction history
      description: |
        Retrieves recent transactions for a token with detailed analysis.
        
        **Current Implementation**: ✅ Complete
        - Transaction signatures and timestamps
        - Transfer patterns
        - Account activity analysis
        - Suspicious activity detection
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
          description: Solana token address
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 100
          description: Number of transactions to retrieve
      responses:
        '200':
          description: Transaction data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/TransactionAnalysis'
                  timestamp:
                    type: string
                    format: date-time
...
components:
  schemas:
...
    TokenAnalysis:
      type: object
      properties:
        tokenAddress:
          type: string
          description: Solana token address
        timestamp:
          type: string
          format: date-time
        tokenInfo:
          $ref: '#/components/schemas/TokenMetadata'
        mintInfo:
          $ref: '#/components/schemas/MintInfo'
        transactionData:
          $ref: '#/components/schemas/TransactionAnalysis'
        creatorData:
          $ref: '#/components/schemas/CreatorAnalysis'
        isWhitelisted:
          type: boolean
          description: Whether token is in known safe list
        whitelistInfo:
          $ref: '#/components/schemas/WhitelistInfo'
        riskScore:
          type: number
          minimum: 0
          maximum: 100
          description: Overall risk score (0-100)
        riskLevel:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        riskFactors:
          type: array
          items:
            type: string
          description: List of identified risk factors
        safetyFactors:
          type: array
          items:
            type: string
          description: List of positive safety indicators
        recommendations:
          type: array
          items:
            type: string
          description: Actionable recommendations
    TokenMetadata:
      type: object
      properties:
        address:
          type: string
        name:
          type: string
        symbol:
          type: string
        decimals:
          type: number
        supply:
          type: string
        mintAuthority:
          type: string
        freezeAuthority:
          type: string
        description:
          type: string
        image:
          type: string
        externalUrl:
          type: string
        interface:
          type: string
        mutable:
          type: boolean
        priceInfo:
          type: object
          properties:
            price_per_token:
              type: number
            currency:
              type: string
    MintInfo:
      type: object
      properties:
        decimals:
          type: number
        supply:
          type: string
        actualSupply:
          type: number
        mintAuthority:
          type: string
        freezeAuthority:
          type: string
        mintRevoked:
          type: boolean
        freezeRevoked:
          type: boolean
    TransactionAnalysis:
      type: object
      properties:
        totalTransactions:
          type: number
        timeSpan:
          type: object
          properties:
            firstTx:
              type: number
            lastTx:
              type: number
            daysActive:
              type: number
        transferPatterns:
          type: object
          properties:
            totalTransfers:
              type: number
            uniqueSenders:
              type: number
            uniqueReceivers:
              type: number
            averageTransferSize:
              type: number
        accountActivity:
          type: object
          properties:
            activeAccounts:
              type: number
            newAccounts:
              type: number
            dormantAccounts:
              type: number
        suspiciousActivity:
          type: object
          properties:
            detected:
              type: boolean
            patterns:
              type: array
              items:
                type: string
            riskLevel:
              type: string
              enum: [LOW, MEDIUM, HIGH]
    CreatorAnalysis:
      type: object
      properties:
        creatorAddress:
          type: string
        creatorActivity:
          type: object
          properties:
            totalTransactions:
              type: number
            lastActivity:
              type: number
            daysSinceLastActivity:
              type: number
        behaviorPatterns:
          type: object
          properties:
            isActive:
              type: boolean
            hasMultipleTokens:
              type: boolean
            suspiciousPatterns:
              type: array
              items:
                type: string
        riskAssessment:
          type: object
          properties:
            level:
              type: string
              enum: [LOW, MEDIUM, HIGH]
            factors:
              type: array
              items:
                type: string
    WhitelistInfo:
      type: object
      properties:
        name:
          type: string
        symbol:
          type: string
        reason:
          type: string
        riskOverride:
          type: string
          enum: [LOW, MEDIUM, HIGH]

    # ========================================
    # DEX ANALYSIS SCHEMAS
    # ========================================
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        timestamp:
          type: string
          format: date-time
    RateLimitError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Too many requests"
        retryAfter:
          type: number
          description: Seconds to wait before retrying

# ========================================
# SECURITY SCHEMAS
# ========================================
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (if required)
security:
  - ApiKeyAuth: []

# ========================================
# ADDITIONAL INFORMATION
# ========================================

# Rate Limiting
# - 100 requests per 15 minutes per IP
# - WebSocket connections are not rate limited
# - Analysis endpoints have higher limits for authenticated users

# WebSocket Events
# - 'analysis:start' - Emitted when analysis begins
#   Data: { tokenAddress: string }
# - 'analysis:complete' - Emitted when analysis completes
#   Data: { tokenAddress: string, riskScore: number, riskLevel: string }
# - 'subscribe:analysis' - Client subscribes to analysis updates
#   Data: { tokenAddress: string }

# Environment Variables Required
# - HELIUS_API_KEY: Required for blockchain data
# - OPENAI_API_KEY: Required for AI analysis (Phase 3)
# - REDIS_URL: Optional for caching
# - CORS_ORIGIN: Frontend origin for CORS
# - PORT: Server port (default: 3000)
# - NODE_ENV: Environment (development/production)
# - LOG_LEVEL: Logging level (debug/info/warn/error) 